requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
openpyxl
python-docx
tensorflow
gradio
